+++ src/app/app-routing.module.ts
@@ -4,9 +4,9 @@ import { LoginMainComponent } from './login-main/login-main.component';
 import { RegisterMainComponent } from './register-main/register-main.component';
 import { ViewMainComponent } from './view-main/view-main.component';
 import { JiraUsersConfigComponent } from './jira-users-config/jira-users-config.component';
-import { RouteGuardsGuard as AuthGuard } from './route-guards.guard';
+import { RouteGuardsGuard as AuthGuard } from './guards/route-guards.guard';
 import { NotfoundComponent } from './notfound/notfound.component';
-import { CantAccessGuard } from './cant-access.guard';
+import { CantAccessGuard } from './guards/cant-access.guard';
 
 const routes: Routes = [
   { path: 'login', component: LoginMainComponent, canActivate: [CantAccessGuard] },
