+++ src/app/app.module.ts
@@ -8,13 +8,12 @@ import { LoginMainComponent } from './login-main/login-main.component';
 import { ViewMainComponent } from './view-main/view-main.component';
 import { NavbarComponent } from './navbar/navbar.component';
 import { FormsModule } from '@angular/forms';
-import { DataManagerService } from './data-manager.service';
-import { ApiService } from './api.service';
+import { ApiService } from './services/api.service';
 import { HttpClientModule } from '@angular/common/http';
 import { JiraUsersConfigComponent } from './jira-users-config/jira-users-config.component';
 import { ShowCertificatesComponent } from './show-certificates/show-certificates.component';
 import { NotfoundComponent } from './notfound/notfound.component';
-import { AuthenticatedService } from './authenticated.service';
+import { AuthenticatedService } from './services/authenticated.service';
 import { NavbarLoginComponent } from './navbar-login/navbar-login.component';
 
 @NgModule({
@@ -36,7 +35,6 @@ import { NavbarLoginComponent } from './navbar-login/navbar-login.component';
     HttpClientModule
   ],
   providers: [
-    DataManagerService,
     ApiService,
     AuthenticatedService
   ],
